added close session button, based on logout button
