+++ public/locales/en-US/translations.json
@@ -797,7 +797,9 @@
     "required": "REQUIRED",
     "proceed": "Proceed",
     "saveAndLaunch": "Save & Launch",
-    "launchNoSave": "Launch without saving"
+    "launchNoSave": "Launch without saving",
+    "logout": "Logout",
+    "closeSession": "Close Session"
   },
   "crashHandler": {
     "text1": "An error occurred that caused the Bitfinex Honey UI to halt. Please, restart the application to proceed working with it",
