+++ src/components/Navbar/Navbar.js
@@ -12,8 +12,7 @@ import UIActions from '../../redux/actions/ui'
 import NavbarLink from './Navbar.Link'
 import NavbarButton from './Navbar.Button'
 import SwitchMode from '../SwitchMode'
-import Logout from './Navbar.Logout'
-
+import CloseSessionButton from './Navbar.CloseSessionButton'
 import LayoutSettings from './Navbar.LayoutSettings'
 import AppSettings from './Navbar.AppSettings'
 import Routes, { strategyEditor } from '../../constants/routes'
@@ -21,9 +20,9 @@ import { isElectronApp } from '../../redux/config'
 import {
   getThemeSetting, THEMES, getIsPaperTrading, getIsBetaVersion, getIsStrategiesTabVisible,
 } from '../../redux/selectors/ui'
+import { UI_MODAL_KEYS } from '../../redux/constants/modals'
 
 import './style.css'
-import { UI_MODAL_KEYS } from '../../redux/constants/modals'
 
 const getOption = (label, url) => (
   <a key={label} className='dropdown-leaf-link' href={url} target='_blank' rel='noopener noreferrer'>
@@ -69,7 +68,7 @@ const Navbar = () => {
         })}
       </ul>
       <div className='hfui-tradingpage__menu'>
-        <div className={cx('hfui-exchangeinfobar__buttons', { 'is-web': !isElectronApp })}>
+        <div className='hfui-exchangeinfobar__buttons'>
           <LayoutSettings />
           <NavbarButton
             alt={t('notifications.title')}
@@ -84,9 +83,6 @@ const Navbar = () => {
             options={leafOptions}
             className='simpledropdown-wrapper'
           />
-          {!isElectronApp && (
-            <Logout />
-          )}
         </div>
         {isElectronApp && (
           <div className='hfui-tradingpaper__control'>
@@ -96,6 +92,7 @@ const Navbar = () => {
             </div>
           </div>
         )}
+        <CloseSessionButton />
       </div>
     </div>
   )
