+++ src/components/Navbar/style.scss
@@ -268,10 +268,6 @@ body.hosted {
 .hfui-exchangeinfobar__buttons {
   display: flex;
   margin-right: spacing();
-
-  &.is-web {
-    margin-right: 0;
-  }
 }
 
 .hfui-exchangeinfobar__button {
@@ -335,9 +331,10 @@ body.hosted {
   }
 }
 
-.hfui-navbar__logout {
+.hfui-navbar__close-session {
   width: auto;
   font-weight: 500;
   padding: 0 spacing(0.75);
   background: none;
+  border: 1px solid var(--text-color);
 }
